% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_additor.R, R/add_extractor.R
\name{add_extractor}
\alias{add_extractor}
\alias{add_extractor}
\title{Add or replace additor}
\usage{
add_extractor(fun)

add_extractor(fun)
}
\arguments{
\item{fun}{function which is run during addittion of the data to the database.}

\item{fun}{function which is run during extracting of the data.}
}
\value{
invisiby the db_additor queue. A \code{list} which is processed

invisiby the extractor queue. A \code{list} which is processed
}
\description{
Adding a named function to the queue of db_additors. If the named function already exists will it be replaced.

Adding a named function to the queue of extractors. If the named function already exists will it be replaced.
}
\details{
The functions must not require any arguments.

The functions must not require any arguments.
}
