% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_data.R
\name{add_new_data}
\alias{add_new_data}
\title{Write a raw data table}
\usage{
add_new_data(create_new_table = FALSE)
}
\arguments{
\item{create_new_table}{if \code{TRUE}, create a new table if it does not exist. Default \code{FALSE}, i.e. raises error if the table does not exist.}
}
\value{
returns invisible\code{TRUE}
}
\description{
Add new data to the database. If a table exist, append it, if not, create a new table with the name \code{name} if \code{create_new_table = TRUE}.
}
