% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pre_processor.R
\name{add_pre_processor}
\alias{add_pre_processor}
\title{Add or replace pre-processor}
\usage{
add_pre_processor(fun)
}
\arguments{
\item{fun}{function which is run during pre-processing of the data.}
}
\value{
invisiby the pre-processor queue. A \code{list} which is processed
}
\description{
Adding a named function to the queue of pre-processors. If the named function already exists will it be replaced.
}
