% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_new_data.R
\name{archive_new_data}
\alias{archive_new_data}
\title{Compress and checksum the \code{get_option("new_data_path")}}
\usage{
archive_new_data(compression = get_option("config")$archive_compression,
  overwrite = FALSE)
}
\arguments{
\item{compression}{which compression should be used for the archive. Thge following values are supported at the moment:
\describe{
   \item{\code{"none"}}{copy the to be imported data folder into archive}
   \item{\code{"tar"}}{tar the to be imported data folder}
   \item{\code{"tar.gz"}}{tar and gz compress the to be imported data folder}
}}

\item{overwrite}{if \code{TRUE}, overwrite existing tar file; default is \code{FALSE}}
}
\value{
invisibly returns the name of the archivefile
}
\description{
Compress all files in \code{get_option("new_data_path")} into the directory \code{"NEW_DATA_PATH/../archive"}
}
