% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemovi_extractor.R
\name{bemovi_extractor}
\alias{bemovi_extractor}
\title{Extractor bemovi data}
\usage{
bemovi_extractor()
}
\value{

}
\description{
Analyse all \code{.avi} files in \code{bemovi} folder and save as \code{.rds} file.
}
\details{
This function is extracting data to be added to the database (and therefore make accessible for further analysis and forecasting)
from \code{.avi} files.
}
