% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemovi_pre_processor.R
\name{bemovi_pre_processor}
\alias{bemovi_pre_processor}
\title{Preprocessor bemovi data}
\usage{
bemovi_pre_processor()
}
\value{

}
\description{
convert all \code{.cxd} files in \code{bemovi} folder to non-proprietory avi format and delete \code{.cxd} file.
}
