% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_new_data.R
\name{check_new_data}
\alias{check_new_data}
\title{Check new data}
\usage{
check_new_data(...)
}
\arguments{
\item{...}{aditional arguments for the test function \code{testthat::test_dir}}
}
\value{
invisibly returns list of two elements as class \code{data_test_results}
\describe{
   \item{\code{result}}{  \code{TRUE} if data is conform with rules, \code{FALSE} if not}
   \item{\code{details}}{ obect of class \code{testthat_results} containing the detailed results of the tests}
}
}
\description{
ONLY DUMMY
Check the data in the directory \code{to_be_imported} for conformity with database rules
}
\examples{

# To check data in the default directory, you can use
check_new_data()
#
# You can also specify a directory in which the data can be found
\dontrun{
check_new_data("./datadir")
}
}
