% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_raw_data.R
\name{connect_raw_data}
\alias{connect_raw_data}
\title{Connect to raw data source}
\usage{
connect_raw_data()
}
\value{
invisibly the connection
}
\description{
The connection is saved in the option \code{raw_data_connection}
}
