% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{DATA_options}
\alias{DATA_options}
\title{Set or get options for this package}
\usage{
DATA_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for this package
}
\section{Supported options}{

The following options are supported TODO
\itemize{
 \item{\code{api_url}} {(\code{character} = \code{"https://api.originstamp.org/api/"}) The url of the OriginStamp API }
 \item{\code{api_key}} {(\code{numeric} = \code{"Please get valid API key from Originstamp.org!"}) The api key.  This needs to be obtained from \url{https://originstamp.org/} }
}
}

\examples{
  ROriginStamp_options("api_key")
  # [1] "Please get valid API key from Originstamp.org!"

  ROriginStamp_options( api_key = "My new API key")
  ROriginStamp_options("api_key")
  # [1] "My new API key"
}
