% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect_data.R
\name{db_connect_data}
\alias{db_connect_data}
\title{Connect to raw data source}
\usage{
db_connect_data()
}
\value{
invisibly the connection
}
\description{
The connection is saved in the option \code{data_connection}
}
