% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_disconnect_raw_data.R
\name{db_disconnect_raw_data}
\alias{db_disconnect_raw_data}
\title{Disconnect from connected raw data source}
\usage{
db_disconnect_raw_data()
}
\value{
result from \code{DBI::dbDisconnect(conn)}
}
\description{
Disconnect from connected raw data source
}
