% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_isAlive_raw_data.R
\name{db_isAlive_data}
\alias{db_isAlive_data}
\title{Check if coinnection to data source is alive}
\usage{
db_isAlive_data()
}
\value{
\code{TRUE} id connection alive, otherwise \code{FALSE}
}
\description{
Checks if connection is alive and functional. If not, the function tries
to close the connection.
}
