% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_hash_new_data.R
\name{delete_hash_new_data}
\alias{delete_hash_new_data}
\title{Delete all hash files}
\usage{
delete_hash_new_data()
}
\value{
invisibly the result from \code{unlink}
}
\description{
Deleta all \code{*.sha256} files in the \code{DATA_options("to_be_imported")} folder recursively.
}
