% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_new_data.R
\name{delete_new_data}
\alias{delete_new_data}
\title{Delete new data}
\usage{
delete_new_data(all = FALSE)
}
\arguments{
\item{all}{if \code{TRUE}, delete \bold{all} files from the}
}
\value{
invisibly the result from \code{unlink(...)}
}
\description{
Delete new data
}
