% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_raw_data.R
\name{disconnect_raw_data}
\alias{disconnect_raw_data}
\title{Disconnect from connected raw data source}
\usage{
disconnect_raw_data()
}
\value{
result from \code{DBI::dbDisconnect(conn)}
}
\description{
Disconnect from connected raw data source
}
