% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_new_data.R
\name{extract_new_data}
\alias{extract_new_data}
\title{Prepare data for import into database}
\usage{
extract_new_data()
}
\value{

}
\description{
Run all the extractors registered with \code{extractors}.
}
\details{
They are saved in a list in the option \code{add_extractor} and processed in order.
}
