% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_flowcytometer.R
\name{extractor_flowcytometer}
\alias{extractor_flowcytometer}
\title{Preprocessor flowcytometer data}
\usage{
extractor_flowcytometer()
}
\value{

}
\description{
Convert all \code{.fxs} files in \code{flowcytometrie} folder to \code{data.frame} and save as \code{.rds} file.
}
\details{
This function is extracting data to be added to the database (and therefore make accessible for further analysis and forecasting)
from \code{.fcs} files.
}
