% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcam_pre_processor.R
\name{flowcam_pre_processor}
\alias{flowcam_pre_processor}
\title{Preprocessor flowcam data}
\usage{
flowcam_pre_processor()
}
\value{

}
\description{
Convert all \code{.tif} files in \code{flowcam} folder to zip compressed TIFF.
}
