% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcytometer_extractor.R
\name{flowcytometer_extractor}
\alias{flowcytometer_extractor}
\title{Preprocessor flowcytometer data}
\usage{
flowcytometer_extractor()
}
\value{

}
\description{
Convert all \code{.fxs} files in \code{flowcytometrie} folder to \code{data.frame} and save as \code{.rds} file.
}
\details{
This function is extracting data to be added to the database (and therefore make accessible for further analysis and forecasting)
from \code{.fcs} files.
}
