% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcytometer_pre_processor.R
\name{flowcytometer_pre_processor}
\alias{flowcytometer_pre_processor}
\title{Preprocessor flowcytometer data}
\usage{
flowcytometer_pre_processor()
}
\value{

}
\description{
Convert all \code{.c6} files in \code{flowcytometrie} folder to \code{.fcs} files and delete \code{.c6} file..
}
