% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_new_data.R
\name{hash_new_data}
\alias{hash_new_data}
\title{Create hash file of new_data}
\usage{
hash_new_data(tts = FALSE, overwrite = FALSE)
}
\arguments{
\item{tts}{if \code{TRUE}, request Trusted Time Stamp for \bold{directory
hash} from OriginStamp and results saved to \code{TTS.result.yml} and
download the seed into the file \code{TTS.seed}; default \code{FALSE}}

\item{overwrite}{if \code{TRUE}, the files \code{file.sha256} and
\code{dir.sha256} will be re-generated and overwritten! default is
\code{FALSE}. Only one ore none of these files exist, both will be
re-generated.}
}
\value{
invisibly \code{TRUE}
}
\description{
This function is calculating the \bold{file hash} for each subdirectory and
file and storing it in a \code{files.sha256} file and finally calculates the
\bold{directory hash} for this file and saves it in the file
\code{dir.sh256}. It is calculating one checksum file for the whole dataset
as well which is stored in the ToBeImported directory.
}
