% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_new_data.R
\name{import_new_data}
\alias{import_new_data}
\title{Title}
\usage{
import_new_data(create_new_table = FALSE)
}
\arguments{
\item{create_new_table}{if \code{TRUE}, create a new table if it does not exist. Default \code{FALSE}, i.e. raises error if the table does not exist.}
}
\value{
\code{TRUE} if import succedded. If failed, report with why it failed TODO
}
\description{
Title
}
