% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incubatortemp_extractor.R
\name{incubatortemp_extractor}
\alias{incubatortemp_extractor}
\title{Extractor incubatortemp data}
\usage{
incubatortemp_extractor()
}
\value{

}
\description{
Convert all \code{.cvs} files in \code{incubatortemp} folder to \code{data.frame} and save as \code{.rds} file.
}
\details{
This function is extracting data to be added to the database (and therefore make accessible for further analysis and forecasting)
from \code{.csv} files.
}
