% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_folder.R
\name{init_folder}
\alias{init_folder}
\title{Create folder structure for data import and processing}
\usage{
init_folder()
}
\value{
invisible \code{TRUE}
}
\description{
Create folder structure for data import and processing
}
