% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_DB.R
\name{initialize_DB}
\alias{initialize_DB}
\title{Create folder structure for data import and processing}
\usage{
initialize_DB()
}
\value{
invisible \code{TRUE}
}
\description{
Create folder structure for data import and processing
}
