% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_register_packages.R
\name{install_register_packages}
\alias{install_register_packages}
\title{Install packages from config file and register the functions to be used}
\usage{
install_register_packages(packages)
}
\arguments{
\item{list}{of packages. Each element \bold{must} contain the elements
\describe{
  \item{\code{name}}{the name of the package,}
  \item{\code{InstallCommand}}{the command to be executed to install the package, and }
  \item{\code{RegisterCommand}}{the command to be executed to register the functions in a queue}
}}
}
\value{
invisibly a list containing the results of the register commands
}
\description{
Install packages from config file and register the functions to be used
}
\examples{
\dontrun{
 install_register_packages(options()$LEEF.Data$measurement_packages)
}
}
