% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_installed_packages.R
\name{LEEF_installed_packages}
\alias{LEEF_installed_packages}
\title{List packages on which LEEF depends}
\usage{
LEEF_installed_packages(recursive = FALSE)
}
\arguments{
\item{recursive}{logical: should (reverse) dependencies of (reverse) dependencies (and so on) be included?}
}
\value{

}
\description{
This function is a wrapper around \code{tools::package_dependencies("LEEF", which = "all")}
}
\examples{
LEEF_installed_packages()
}
