% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualcount_extractor.R
\name{manualcount_extractor}
\alias{manualcount_extractor}
\title{Extractor manualcount data}
\usage{
manualcount_extractor()
}
\value{

}
\description{
Convert all \code{.cvs} files in \code{manualcount} folder to \code{data.frame} and save as \code{.rds} file.
}
\details{
This function is extracting data to be added to the database (and therefore make accessible for further analysis and forecasting)
from \code{.csv} files.
}
