% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_ok.R
\name{new_data_ok}
\alias{new_data_ok}
\title{Check new data}
\usage{
new_data_ok(new_data_dir = file.path(get_option("pkg_dir", "new_data")), ...)
}
\arguments{
\item{...}{aditional arguments for the test function \code{testthat::test_dir}}
}
\value{
invisibly returns list of two elements as class \code{data_test_results}
\describe{
   \item{\code{result}:}{  \code{TRUE} if data is conform with rules, \code{FALSE} if not}
   \item{\code{details}:}{ obect of class \code{testthat_results} containing the detailed results of the tests}
}
}
\description{
Check the data in the directory \code{new_data_dir} for conformity with database rules
}
\examples{

# To check data in the default directory, you can use
new_data_ok()
#
# You can also specify a directory in which the data can be found
\dontrun{
new_data_ok("./datadir")
}
}
