% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{opt_directories}
\alias{opt_directories}
\title{Functions to read and write options}
\usage{
opt_directories(raw, pre_processed, extracted, archive, tools)
}
\arguments{
\item{raw}{\code{character} \code{vector} of length one containing the
directory for the raw data}

\item{pre_processed}{\code{character} \code{vector} of length one containing
the directory for the pre_processed data}

\item{extracted}{\code{character} \code{vector} of length one containing the
directory for the extracted data}

\item{archive}{\code{character} \code{vector} of length one containing the
directory for the archived data}

\item{tools}{directory in which the tools are located}
}
\value{
list of directories. If values have set, the value before the change.
}
\description{
Read or write the directories to be used in the processing. Directories do
not have to exist and will be created. Content will be overwritten without
confirmation! If no parameter is given, the directories will be returned a a list.
}
\examples{
opt_directories()

opt_directories(raw = "./temp")

}
