% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_bemovi.R
\name{pre_process_bemovi}
\alias{pre_process_bemovi}
\title{Preprocessor bemovi data}
\usage{
pre_process_bemovi()
}
\value{

}
\description{
convert all \code{.cxd} files in \code{bemovi} folder to non-proprietory avi format
}
