% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_flowcam.R
\name{pre_process_flowcam}
\alias{pre_process_flowcam}
\title{Preprocessor flowcam data}
\usage{
pre_process_flowcam()
}
\value{

}
\description{
Convert all \code{.tif} files in \code{flowcam} folder to zip compressed TIFF.
}
