% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_flowcytometer.R
\name{pre_process_flowcytometer}
\alias{pre_process_flowcytometer}
\title{Preprocessor flowcytometer data}
\usage{
pre_process_flowcytometer()
}
\value{

}
\description{
Convert all \code{.c6} files in \code{flowcytometrie} folder to \code{.fcs} files and delete \code{.c6} file..
}
