% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_new_data.R
\name{pre_process_new_data}
\alias{pre_process_new_data}
\title{Preprocessor for the different data sources}
\usage{
pre_process_new_data()
}
\value{

}
\description{
Is used for example for conversion of uncompressed TIFF to compressed TIFF. Is executed \bold{before} hashing and archiving and after checking!.
}
