% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Process all ques in the correct order}
\usage{
process(submitter, timestamp, process = TRUE, ...)
}
\arguments{
\item{submitter}{name of submitter. When provided, will override the one in
the `sample_metadata.yml` file.}

\item{timestamp}{timestamp for the data. When provided, will override the one in
the `sample_metadata.yml` file.}

\item{process}{if \code{TRUE}, the pipeline will be processed. if
\code{FALSE}, only the checks of the config file will be done nd no actual
processing is happening.}

\item{...}{additional arguments for the different queues}
}
\value{
invisibly \code{TRUE}
}
\description{
This function is an example and can be used as a template for processing the
queues in a script,. Raw data is always archived using the "none"
compression.
}
\examples{
\dontrun{
 process()
}
}
