% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_comp_none.R
\name{process_raw_comp_none}
\alias{process_raw_comp_none}
\title{Process all ques in the correct order}
\usage{
process_raw_comp_none(submitter, timestamp, ...)
}
\arguments{
\item{submitter}{name of the submitter of the data to the pipeline. Will be added to the metadata.}

\item{timestamp}{timestamp of the submission of the data to the pipeline.
This should be in the format \code{YYYYMMDD} and will be used to identify the sampling day.}

\item{...}{additional arguments for the different queues}
}
\value{
invisibly \code{TRUE}
}
\description{
This function is an example and can be used as a template for processing the
queues in a script. It uses the archiver "none" for the raw and pre-processed
data, useful for already compressed and large data, e.g. bemovi.
}
\examples{
\dontrun{
 process()
}
}
