% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_new_data_hash.R
\name{read_new_data_hash}
\alias{read_new_data_hash}
\title{Read hash from new data file}
\usage{
read_new_data_hash(file)
}
\arguments{
\item{file}{if given, function returns the hash of the file, if not, a
dataframe containing the hashes}
}
\value{
the hash of the file (if given), otherwise a \code{data.frame}
  containing the columns \code{file} containing the file names and
  \code{hash} containing the hashes. The row names are the file names as
  well.
}
\description{
Function to read the hash of a new data file. The function looks for the file
\code{hash.sha256} in the \code{get_option("new_data_path")} directory.
}
