% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_packages.R
\name{register_packages}
\alias{register_packages}
\title{Register the functions to be usedfrom packages in the config file}
\usage{
register_packages(packages)
}
\arguments{
\item{list}{of packages. Each element \bold{must} contain the elements
\describe{
  \item{\code{name}}{the name of the package,}
  \item{\code{InstallCommand}}{the command to be executed to install the package, and }
  \item{\code{RegisterCommand}}{the command to be executed to register the functions in a queue}
}}
}
\value{
invisibly a list containing the results of the register commands
}
\description{
Register the functions to be usedfrom packages in the config file
}
\examples{
\dontrun{
 register_packages(getOption("LEEF.Data")$measurement_packages)
}
}
