% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respirometer_extractor.R
\name{respirometer_extractor}
\alias{respirometer_extractor}
\title{Extractor respirometer data}
\usage{
respirometer_extractor()
}
\value{

}
\description{
Convert all \code{.cvs} files in \code{respirometer} folder to \code{data.frame} and save as \code{.rds} file.
}
\details{
This function is extracting data to be added to the database (and therefore make accessible for further analysis and forecasting)
from \code{.csv} files.
}
