% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run process queue}
\usage{
run(input, output, queue)
}
\arguments{
\item{input}{directory containing the input data in folders with the name
of the methodology (e.g. \code{bemovi})}

\item{output}{directory in which the results will be written in a folder with
the name of the methodology (e.g. \code{bemovi})}

\item{queue}{name of queue in \code{getOption("LEEF")}}
}
\value{
returns the results of the queue as a vector of length of the queue.
  If an element is \code{TRUE}, the function was run successfully (i.e.
  returned \code{TRUE})
}
\description{
Run all the functions in the process queue named \code{queue}
}
\examples{
\dontrun{
run(
  input = "./input",
  output = "./output",
  queue = "extractor"
)
}
}
