% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_additors.R
\name{run_additors}
\alias{run_additors}
\title{Run additors queue}
\usage{
run_additors()
}
\value{

}
\description{
Run all the additors registered with \code{add_additor()}.
}
\examples{
\dontrun{
run_additors()
}
}
