% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_archivers.R
\name{run_archivers}
\alias{run_archivers}
\title{Run archivers queue}
\usage{
run_archivers(input, output)
}
\arguments{
\item{input}{directory to be archive, including subdirectories}

\item{output}{director in which the archive will be created}
}
\value{
returns the results of the queue as a vector of length of the queue.
  If an element is \code{TRUE}, the function was run successfully (i.e.
  returned \code{TRUE})
}
\description{
Run all the archivers registered with \code{add_archiver()}.
}
\examples{
\dontrun{
run_archivers(
  input = "./input",
  output = "./output"
)
}
}
