% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pre_processor.R
\name{run_pre_processors}
\alias{run_pre_processors}
\title{Run pre_processors queue}
\usage{
run_pre_processors()
}
\value{

}
\description{
Run all the additors registered with \code{add_pre_processor()}.
}
\examples{
\dontrun{
run_pre_processors()
}
}
