% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_bemovi.R
\name{split_bemovi}
\alias{split_bemovi}
\title{Split bemovi filder into a number of bemovi. folders with a maximum of
per_batch video files}
\usage{
split_bemovi(
  per_batch = 30,
  bemovi_dir = file.path(".", "0.raw.data"),
  overwrite = TRUE
)
}
\arguments{
\item{per_batch}{maximum number of movies per batch}

\item{bemovi_dir}{bas directory in which the \code{bemovi} directory is
located}

\item{overwrite}{if \code{TRUE}, all folders starting with \code{bemovi.} in
the \code{bemovi_dir} will be deleted}
}
\value{
the maximum id used
}
\description{
Split bemovi filder into a number of bemovi. folders with a maximum of
per_batch video files
}
\examples{
\dontrun{
split_bemovi(per_batch = 5)
}
}
