% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_new_data.R
\name{tar_new_data}
\alias{tar_new_data}
\title{Compress and checksum the \code{get_option("new_data_path)}}
\usage{
tar_new_data()
}
\value{
invisibly returns \code(TRUE)
}
\description{
Compress all files in \code{get_option("new_data_path")} into the directory \code{"NEW_DATA_PATH/../new_data_archive}
}
