% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_raw_data.R
\name{write_raw_data}
\alias{write_raw_data}
\title{Write a raw data table}
\usage{
write_raw_data(name, value)
}
\arguments{
\item{name}{name of table in database}

\item{value}{tabel to be written to database}
}
\value{
returns invisible the rerturn valuer of \code{DBI::WriteTable()}
}
\description{
Write \code{x} to a table in the database. If the table exist, append it, if not, create a new table with the name \code{name}.
}
