% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_extractor.R
\name{add_extractor}
\alias{add_extractor}
\title{Add or replace extractor}
\usage{
add_extractor(fun)
}
\arguments{
\item{fun}{function which is run during extracting of the data.}
}
\value{
invisiby the extractor queue. A \code{list} which is processed
}
\description{
Adding a named function to the queue of extractors. If the named function already exists will it be replaced.
}
\details{
The functions must not require any arguments.
}
