% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcytometer_extractor.R
\name{flowcytometer_extractor}
\alias{flowcytometer_extractor}
\title{Preprocessor flowcytometer data}
\usage{
flowcytometer_extractor()
}
\value{

}
\description{
Convert all \code{.fxs} files in \code{flowcytometrie} folder to \code{data.frame} and save as \code{.rds} file.
}
