% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_new_data.R
\name{hash_new_data}
\alias{hash_new_data}
\title{Create hash file of new_data}
\usage{
hash_new_data(tts = FALSE, overwrite = FALSE)
}
\arguments{
\item{tts}{if \code{TRUE}, request Trusted Time Stamp for directory hash from
OriginStamp and results saved to \code{TTS.yml}; default \code{FALSE}}

\item{overwrite}{if \code{TRUE}, the files \code{file.sha256.txt} and
\code{dir.sha256.txt} will be re-generated and overwritten! default is
\code{FALSE}. Only one ore none of these files exist, both will be
re-generated.}
}
\value{
invisibly \code{TRUE}
}
\description{
This function is calculating the \bold{file hash} for each file and storing
it in a \code{files.sha256.txt} file and finally calculates the
\bold{directory hash} for this file and saves it in the file
\code{dir.sh256.txt}
}
