% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Create folder structure for data import and processing}
\usage{
initialize(config_file = system.file("default_config.yml", package = "LEEF"))
}
\arguments{
\item{config_file}{config file to use. If none is specified, \code{cofig.yml} in the current working directory will be used.}
}
\value{
invisible \code{TRUE}
}
\description{
Create folder structure for data import and processing
}
\examples{
\dontrun{
initialize(system.file("default_config.yml", package = "LEEF"))
}
}
