% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_db.R
\name{initialize_db}
\alias{initialize_db}
\title{Create folder structure for data import and processing}
\usage{
initialize_db()
}
\value{
invisible \code{TRUE}
}
\description{
Create folder structure for data import and processing
}
