% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_DB.R
\name{initialize_db}
\alias{initialize_db}
\title{Create folder structure for data import and processing}
\usage{
initialize_db(config_file)
}
\arguments{
\item{config_file}{config file to use. If none is specified, \code{cofig.yml} in the current working directory will be used.}
}
\value{
invisible \code{TRUE}
}
\description{
Create folder structure for data import and processing
}
