% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_LEEF_packages.R
\name{list_LEEF_packages}
\alias{list_LEEF_packages}
\title{List packages on which LEEF depends from the LEEF-UZH repo}
\usage{
list_LEEF_packages(recursive = TRUE, versions = FALSE)
}
\arguments{
\item{recursive}{logical: should (reverse) dependencies of (reverse) dependencies (and so on) be included? defaults to \code{TRUE}}

\item{versions}{logical: should versions be returned as well.}
}
\value{
list of all packages which are installed which contain \bold{.LEEF} or \bold{LEEF.} and the package \bold{LEEF} itself
}
\description{
This function is a wrapper around \code{tools::package_dependencies("LEEF",
which = "all", recursive = TRUE)} which returns only the packages which
contain \bold{.LEEF} or \bold{LEEF.} and the package \bold{LEEF} itself.
}
\examples{
\dontrun{
list_LEEF_packages()
}
}
