% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_new_data.R
\name{pre_process_new_data}
\alias{pre_process_new_data}
\title{Preprocessor for the different data sources}
\usage{
pre_process_new_data()
}
\value{

}
\description{
Run all the pre-processors registered with \code{add_pre_processor}.
}
\details{
They are saved in a list in the option \code{pre_processors} and processed in order.
}
