% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Process all ques in the correct order}
\usage{
process(submitter, timestamp, ...)
}
\arguments{
\item{...}{additional arguments for the different queues}
}
\value{
invisibly \code{TRUE}
}
\description{
This function is an example and can be used as a template for processing the queues in a script
}
\examples{
\dontrun{
 process()
}
}
