% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_new_data.R
\name{read_new_data}
\alias{read_new_data}
\title{Read new data file}
\usage{
read_new_data(file)
}
\arguments{
\item{file}{name of the file in \code{get_option("new_data_path)} to be read}
}
\value{
data frame containing the new data
}
\description{
Funcrtion to read a new data file. The function looks for the file in the \code{get_option("new_data_path)}.
It sets values to define the format used to guarantee consistency.
}
