% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_additors.R
\name{run_additors}
\alias{run_additors}
\title{Run additors queue}
\usage{
run_additors()
}
\value{
returns the results of the queue as a vector of length of the queue.
  If an element is \code{TRUE}, the function was run successfully (i.e.
  returned \code{TRUE})
}
\description{
Run all the additors registered with \code{add_additor()}.
}
\examples{
\dontrun{
run_additors()
}
}
