% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_extractor.R
\name{run_extractors}
\alias{run_extractors}
\title{Run extractors queue}
\usage{
run_extractors()
}
\value{

}
\description{
Run all the additors registered with \code{add_extractor()}.
}
\examples{
\dontrun{
run_extractors()
}
}
